refactor(view): move the Details panel out of view.go
Details renders the widest set of fields in the interface and every one of them
is width-aware, which made it the largest theme in the file.

Declarations move unchanged.
